update ModuleCard usb info
